feat: add helper for interaction responses
